[level 1] Title: 평균 구하기, Time: 0.03 ms, Memory: 81.5 MB -BaekjoonHub

+++ 프로그래머스/lv1/12944. 평균 구하기/평균 구하기.java
@@ -1,12 +1,13 @@
 class Solution {
     public double solution(int[] arr) {
-        int add = 0;
+        double add = 0;
         
         for(int i : arr) {
-            add += i;        
+            add += i;
         }
         
-        double answer = (double) add / arr.length;
+        double answer = add / arr.length;
+        
         return answer;
     }
 }
